feat: add and implement `date-converter` module
